+++ src/main.tsx
@@ -1,17 +1,12 @@
-import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { RootProvider } from '@/context/root-provider';
 
 import './styles/global.css';
 import { App } from './pages';
 
-const queryClient = new QueryClient();
-
 createRoot(document.querySelector('#root')!).render(
-  <React.StrictMode>
-    <QueryClientProvider client={queryClient}>
-      <App />
-    </QueryClientProvider>
-  </React.StrictMode>,
+  <RootProvider>
+    <App />
+  </RootProvider>,
 );
