--- src/services/enums/product-type.tsx
@@ -1,5 +1,0 @@
-export enum ProductTypeEnum {
-  Product = 'product',
-  ManufacturingPart = 'manufacturing-part',
-  PurchasingPart = 'purchasing-part',
-}
