--- src/services/types/attribute.tsx
@@ -1,4 +1,0 @@
-export interface IAttribute {
-  name: string;
-  option: string;
-}
