--- src/services/types/value.tsx
@@ -1,7 +1,0 @@
-export interface IValue {
-  name: string;
-  value: number;
-  unit: string;
-  toleranceMin: number;
-  toleranceMax: number;
-}
